Sequences

+++ src/irvine/oeis/a015/A015723.java
@@ -9,13 +9,13 @@ import irvine.oeis.a008.A008289;
  */
 public class A015723 extends A008289 {
 
-  private long mN = 0;
+  private int mN = 0;
 
   @Override
   public Z next() {
     ++mN;
     Z sum = Z.ZERO;
-    for (long k = 1; k <= mN; ++k) {
+    for (int k = 1; k <= mN; ++k) {
       sum = sum.add(get(mN, k).multiply(k));
     }
     return sum;
